chore: move sortResults and update tests
