+++ src/formatters/__tests__/teamcity.test.ts
@@ -1,6 +1,7 @@
+import { sortResults } from '../../utils';
 import { teamcity } from '../teamcity';
 
-const mixedErrors = require('./__fixtures__/mixed-errors.json');
+const mixedErrors = sortResults(require('./__fixtures__/mixed-errors.json'));
 
 describe('Teamcity formatter', () => {
   test('should format messages', () => {
