+++ src/utils/__tests__/sortResults.test.ts
@@ -1,5 +1,5 @@
 import { DiagnosticSeverity } from '@stoplight/types';
-import { IRuleResult } from '../../../types';
+import { IRuleResult } from '../../types';
 import { compareResults, sortResults } from '../sortResults';
 
 const results: IRuleResult[] = [
