+++ src/utils/index.ts
@@ -3,3 +3,4 @@ export * from './hasIntersectingElement';
 export * from './isObject';
 export * from './refs';
 export * from './prepareResults';
+export * from './sortResults';
